refactor(admin): lazy-load codemirror json editor
